feat: only handle SnapshotSent you emitted
tldr people could submit malicious or
spammy sendSnapshot since they're so
cheap so call. as a result, our validator
would get overwhelmed and waste
gas relaying txs that would
revert uselessly anyway.
all thats needed is passing a bogus
Claim

theres a better way of doing this but
its very complicated. a simpler compromise:
since validator will call sendSnapshot
every time it listens to Challenged,
validator will be able to index its
own SnapshotSent events and relay
its own version.

the better way would be something
like, checking the calldata of the
tx thats going to be relayed,
comparing the Claim to the one
in the ClaimStore, and then once its
considered to be the correct one
AND that it wasnt SnapshotSent
previously before, THEN you schedule
the relaying of the tx  in L1

and then, the isSpent check would
be put to use. but thats not the
world we're living in, so we're gonna
go with this for the 1st version